Implementing gun rotation with  L/R arrows Restructure project into mod files Running formatter across changed files

+++ src/wall/mod.rs
@@ -1,11 +1,11 @@
 pub mod models;
 
 use crate::person::models::Person;
-use models::*;
 use bevy::{
     prelude::*,
     sprite::collide_aabb::{collide, Collision},
 };
+use models::*;
 
 pub fn setup_walls(mut commands: Commands) {
     commands.spawn().insert(Wall).insert_bundle(WallBundle {
@@ -35,27 +35,29 @@ pub fn check_wall_collision(
     mut q_person: Query<&mut Transform, (With<Person>, Changed<Transform>)>,
     q_wall: Query<&Transform, (With<Wall>, Without<Person>)>,
 ) {
-    let mut person_transform = q_person.single_mut();
-    for wall_transform in q_wall.iter() {
-        if let Some(collision) = collide(
-            person_transform.translation,
-            person_transform.scale.truncate(),
-            wall_transform.translation,
-            wall_transform.scale.truncate(),
-        ) {
-            let mut x_delta = 0.;
-            let mut y_delta = 0.;
+    let person_transform = q_person.get_single_mut();
+    if let Ok(mut person_transform) = person_transform {
+        for wall_transform in q_wall.iter() {
+            if let Some(collision) = collide(
+                person_transform.translation,
+                person_transform.scale.truncate(),
+                wall_transform.translation,
+                wall_transform.scale.truncate(),
+            ) {
+                let mut x_delta = 0.;
+                let mut y_delta = 0.;
 
-            match collision {
-                Collision::Left => x_delta -= 2.,
-                Collision::Right => x_delta += 2.,
-                Collision::Top => y_delta += 2.,
-                Collision::Bottom => y_delta -= 2.,
-                _ => {}
+                match collision {
+                    Collision::Left => x_delta -= 2.,
+                    Collision::Right => x_delta += 2.,
+                    Collision::Top => y_delta += 2.,
+                    Collision::Bottom => y_delta -= 2.,
+                    _ => {}
+                }
+
+                person_transform.translation.x += x_delta;
+                person_transform.translation.y += y_delta;
             }
-
-            person_transform.translation.x += x_delta;
-            person_transform.translation.y += y_delta;
         }
     }
 }
